refactor: moved integration tests to separate directory
