+++ internal/tests/stubs/currency_api_client.go
