+++ internal/tests/stubs/mailman.go
